Move ResultDisplay class
