+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -17,6 +17,7 @@ import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.ui.body.BodyPanel;
+import seedu.address.ui.result.ResultDisplay;
 
 /**
  * The Main Window. Provides the basic application layout containing
